+++ Forex/src/backend/Forex.Domain/Entities/ResidueManufactory.cs
@@ -4,8 +4,8 @@ using Forex.Domain.Commons;
 
 public class ResidueManufactory : Auditable
 {
-    public int SemiProductId { get; set; }
-    public int ManufactoryId { get; set; }
+    public long SemiProductId { get; set; }
+    public long ManufactoryId { get; set; }
     public decimal Quantity { get; set; }
 
     public SemiProduct SemiProduct { get; set; } = default!;
