+++ Forex/src/backend/Forex.Domain/Entities/ResidueShop.cs
@@ -4,8 +4,8 @@ using Forex.Domain.Commons;
 
 public class ResidueShop : Auditable
 {
-    public int ProductId { get; set; }
-    public int ShopId { get; set; }
+    public long ProductId { get; set; }
+    public long ShopId { get; set; }
     public decimal Quantity { get; set; }
 
     public Product Product { get; set; } = default!;
