+++ Forex/src/backend/Forex.Domain/Entities/SaleItem.cs
@@ -4,8 +4,8 @@ using Forex.Domain.Commons;
 
 public class SaleItem : Auditable
 {
-    public int SaleId { get; set; }
-    public int ProductId { get; set; }
+    public long SaleId { get; set; }
+    public long ProductId { get; set; }
     public int Count { get; set; }
     public decimal CostPrice { get; set; }
     public decimal Benifit { get; set; }
